Add an initial source file and a build
The plan here is start with the bytecode compiler (because that the one I feel less sure I understand
just from reading Crafting Interpreters). Later they may be a tree-walk-interpreter directory as well.

We're using meson to automatically clone and cppfront and have provided a wrap file and a build patch.